refactor(s3reader): split into separate files and move to folder
- Move S3Reader implementation into its own directory
- Create separate files for base ABC class, sequential/range_based readers, and s3reader_config.
- Keep reader classes in __init__ for tests to reference them
